+++ packages/crdt_lf/lib/src/handler/map/operation.dart
@@ -13,6 +13,8 @@ class _MapOperationFactory<T> {
       return _MapInsertOperation<T>.fromPayload(payload);
     } else if (payload['type'] == OperationType.delete(handler).toPayload()) {
       return _MapDeleteOperation<T>.fromPayload(payload);
+    } else if (payload['type'] == OperationType.update(handler).toPayload()) {
+      return _MapUpdateOperation<T>.fromPayload(payload);
     }
 
     return null;
@@ -42,7 +44,7 @@ class _MapInsertOperation<T> extends Operation {
     required String key,
     required T value,
   }) {
-    return _MapInsertOperation(
+    return _MapInsertOperation<T>(
       id: handler.id,
       type: OperationType.insert(handler),
       key: key,
@@ -82,7 +84,7 @@ class _MapDeleteOperation<T> extends Operation {
     Handler<dynamic> handler, {
     required String key,
   }) {
-    return _MapDeleteOperation(
+    return _MapDeleteOperation<T>(
       id: handler.id,
       type: OperationType.delete(handler),
       key: key,
@@ -96,5 +98,46 @@ class _MapDeleteOperation<T> extends Operation {
         'type': type.toPayload(),
         'id': id,
         'key': key,
+      };
+}
+
+class _MapUpdateOperation<T> extends Operation {
+  const _MapUpdateOperation({
+    required this.key,
+    required this.value,
+    required super.id,
+    required super.type,
+  });
+
+  factory _MapUpdateOperation.fromHandler(
+    Handler<dynamic> handler, {
+    required String key,
+    required T value,
+  }) {
+    return _MapUpdateOperation(
+      id: handler.id,
+      type: OperationType.update(handler),
+      key: key,
+      value: value,
+    );
+  }
+
+  factory _MapUpdateOperation.fromPayload(Map<String, dynamic> payload) =>
+      _MapUpdateOperation<T>(
+        id: payload['id'] as String,
+        type: OperationType.fromPayload(payload['type'] as String),
+        key: payload['key'] as String,
+        value: payload['value'] as T,
+      );
+
+  final String key;
+  final T value;
+
+  @override
+  Map<String, dynamic> toPayload() => {
+        'type': type.toPayload(),
+        'id': id,
+        'key': key,
+        'value': value,
       };
 }